Changed FlowMap to use priorities and ID
Fixed a problem that the OM was having with managing FlowSpace.
Manipulating fields by index (both for identifying and sorting
FlowEntries) is error prone with multiple editors and causes
editors to have to keep the index table as extra state.

I've decoupled identifying and sorting into an id and priority
fields.  So now you delete and change using ID and the flowentries
are listed in order by priority.  Entries with equal priority
are sorted by ID.

All regression tests pass.

known bug:
- priority somehow gets lost going up API
- priority should be descending