fix #124: intersection with priorities was broken
When creating a sub flowspace or testing if we should
allow a slice access to a switch, need to use matches()
instead of intersects() for FlowMap.

This is a bit subtle and may likely come back to bite us.

FlowMap.intersects() does not (and cannot because of openflow)
handle priorities correctly in all cases.  Needs more testing.

Added lots of junit tests and a regression test: tests-vlan.py