FLOWVISOR-73: do buffer-id isolation between slices
buffer-id's are a pointer to a buffered packet.  In theory,
a slice should only be able to address packets sent to
it, but existing code had no such protections.

Now we track buffer_id to slice mappings and prevent slices
from accessing other slices buffer ids.  This is of particular
use with read-only slices.

* fixed up all of the regression tests
* added a new regression test